4. RegisterPostDTO 분리
