+++ src/main/java/com/together/buytogether/post/feature/RegisterPost.java
@@ -6,19 +6,15 @@ import com.together.buytogether.member.domain.SessionConst;
 import com.together.buytogether.member.domain.SessionManager;
 import com.together.buytogether.post.domain.Post;
 import com.together.buytogether.post.domain.PostRepository;
-import com.together.buytogether.post.domain.PostStatus;
+import com.together.buytogether.post.dto.request.RegisterPostDTO;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDateTime;
 
 @RestController
 public class RegisterPost {
@@ -38,37 +34,14 @@ public class RegisterPost {
 
     @PostMapping("/posts")
     @ResponseStatus(HttpStatus.CREATED)
-    public void request(@RequestBody @Valid Request request,
+    public void request(@RequestBody @Valid RegisterPostDTO registerPostDTO,
                         HttpServletRequest httpServletRequest) {
         Cookie cookie = httpServletRequest.getCookies()[0];
         Long memberId = (Long) sessionManger.getSession(cookie.getName()).getAttribute(SessionConst.LOGIN_MEMBER);
         Member member = memberRepository.findById(memberId).orElseThrow(
                 () -> new IllegalArgumentException("존재하지 않는 회원입니다"));
-        Post post = request.toDomain(member);
+        Post post = registerPostDTO.toDomain(member);
         postRepository.save(post);
     }
 
-    public record Request(
-            @NotBlank(message = "글 제목은 필수입니다")
-            String title,
-            @NotBlank(message = "글 내용은 필수입니다")
-            String content,
-
-            @NotNull(message = "글 상태는 필수입니다")
-            PostStatus status,
-            @NotNull(message = "글 만료일은 필수입니다")
-            LocalDateTime expiredAt) {
-
-
-        public Post toDomain(Member member) {
-
-            return new Post(
-                    member,
-                    title,
-                    content,
-                    status,
-                    expiredAt
-            );
-        }
-    }
 }
